// src/app/services/engines/rules/rule-type.enum.ts
export enum RuleTypeEnum {
    MinTravel = "com.magiis.magiis.engines.price.rules.MinTravelRule",
    LongTravel = "com.magiis.magiis.engines.price.rules.LongTravelsRule",
    VehicleType = "com.magiis.magiis.engines.price.rules.VehicleTypeRule",
    AdditionalStop = "com.magiis.magiis.engines.price.rules.AdditionalStopRule",
    WaitTime = "com.magiis.magiis.engines.price.rules.WaitTimeRule",
    PaymentMethodDiscount = "com.magiis.magiis.engines.price.rules.PaymentMethodRule",
    MinimalCost = "com.magiis.magiis.engines.price.rules.MinimalCostRule",
    SegmentPrice = "com.magiis.magiis.engines.price.rules.SegmentPriceRule",
    RoundTrip = "com.magiis.magiis.engines.price.rules.RoundTripRule",
    RentalCarRule = "com.magiis.magiis.engines.price.rules.RentalCarRule",
    TravelTimeRule = "com.magiis.magiis.engines.price.rules.TravelTimeRule",
    TaxiingCostRule = "com.magiis.magiis.engines.price.rules.TaxiingCostRule"
}