// src/app/models/one-shot-rules.model.ts
export class OneShotRulesModel {
  amount?: number;
  tolerancePercent?: number;
  extraKmValue?: number;
  includeToll?: boolean;
  includeParking?: boolean;
  includeOther?: boolean;
  includeOvernight?: boolean;
  includedWaitTime?: number;
  extraWaitTimeValue?: number;
  extraWaypointValue?: number;
  travelCostWithMarkup?: number;
  constructor() {}
}